requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
python-dotenv
requests